Fim Enuns_Datas_Valores

+++ Banco/SRC/com/algawork/aula/modelo/pagamento/Boleto.java
@@ -1,21 +1,23 @@
 package com.algawork.aula.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 import com.algawork.aula.modelo.Pessoa;
 
 public class Boleto implements DocumentoPagavel, DocumentoExtornavel {
 
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean pago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
 
     }
 
     @Override
-    public double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valor;
     }
