+++ Banco/SRC/com/algawork/aula/modelo/pagamento/DocumentoPagavel.java
@@ -1,8 +1,10 @@
 package com.algawork.aula.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 public interface DocumentoPagavel {
    
-        double getValorTotal();
+        BigDecimal getValorTotal();
         boolean estaPago();
         void quitarPagamenmto();
         default void imprimirRecibo(){
